requirements.txt
google-api-python-client==2.58.0
youtube-transcript-api==0.5.0
langchain
openai
tqdm==4.64.1
transformers
python-dotenv==1.0.1
torch
langchain_community
langchain_openai
whisper
youtube_dl
yt-dlp
pydub